added voluntary disconnect for stop button
